# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/bets": {
      "post": {
        "operationId": "BettingController_place",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Clé par tentative, réutilisée au retry HTTP (anti double-débit).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceBetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pari posé (cote figée)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceBetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          },
          "409": {
            "description": "Même Idempotency-Key réutilisée avec un corps différent"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bets"
        ]
      },
      "get": {
        "operationId": "BettingController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paris de l'utilisateur authentifié uniquement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BetViewDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bets"
        ]
      }
    },
    "/bets/{id}": {
      "get": {
        "operationId": "BettingController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetViewDto"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          },
          "404": {
            "description": "Pari introuvable (ou non possédé — anti-IDOR)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bets"
        ]
      }
    },
    "/bets/{id}/events": {
      "get": {
        "operationId": "BettingController_events",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timeline du pari (journal append-only)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BetEventDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          },
          "404": {
            "description": "Pari introuvable (ou non possédé — anti-IDOR)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bets"
        ]
      }
    },
    "/markets/settle": {
      "post": {
        "operationId": "SettlementController_settle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleMarketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Règlement effectué (comptes par statut) — MANAGER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettleMarketResultDto"
                }
              }
            }
          },
          "400": {
            "description": "outcomes vide, ou winningOutcomeId manquant sans voided"
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          },
          "403": {
            "description": "Réservé au rôle MANAGER"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "markets"
        ]
      }
    },
...
    "/markets": {
      "get": {
        "operationId": "CatalogController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Marchés ouverts (modèle N-issues) — PUBLIC",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMarketRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marché créé (MANAGER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDto"
                }
              }
            }
          },
          "400": {
            "description": "Corps invalide (name/game/outcomes)"
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          },
          "403": {
            "description": "Réservé au rôle MANAGER"
          },
          "422": {
            "description": "Marché invalide (≥ 2 issues requises)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Corps invalide (username/password/role)"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer requis/invalide"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "BetNext API",
    "description": "Contrat OpenAPI du back BetNext — source du client TYPÉ du front (généré, non écrit à la main).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PlaceBetRequest": {
        "type": "object",
        "properties": {
          "outcomeId": {
            "type": "string",
            "example": "lol-finale-a"
          },
          "stake": {
            "type": "number",
            "example": 20,
            "minimum": 0,
            "description": "Mise (strictement positive)"
          }
        },
        "required": [
          "outcomeId",
          "stake"
        ]
      },
      "PlaceBetResponse": {
        "type": "object",
        "properties": {
          "betId": {
            "type": "string",
            "example": "3f9c0b2e-..."
          },
          "lockedOdds": {
            "type": "number",
            "example": 2,
            "description": "Cote figée à la pose"
          },
          "potentialGain": {
            "type": "number",
            "example": 40
          },
          "pricingProvisional": {
            "type": "boolean",
            "example": true,
            "description": "Cote d'ouverture (read-model froid)"
          }
        },
        "required": [
          "betId",
          "lockedOdds",
          "potentialGain"
        ]
      },
      "BetViewDto": {
        "type": "object",
        "properties": {
          "betId": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "example": "demo-player"
          },
          "outcomeId": {
            "type": "string",
            "example": "lol-finale-a"
          },
          "stake": {
            "type": "number",
            "example": 20
          },
          "lockedOdds": {
            "type": "number",
            "example": 2
          },
          "potentialGain": {
            "type": "number",
            "example": 40
          },
          "status": {
            "type": "string",
            "example": "PENDING",
            "enum": [
              "PENDING",
              "WON",
              "LOST",
              "VOID",
              "COMPENSATING",
              "REFUNDED"
            ]
          }
        },
        "required": [
          "betId",
          "userId",
          "outcomeId",
          "stake",
          "lockedOdds",
          "potentialGain",
          "status"
        ]
      },
      "BetEventDto": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "number",
            "example": 1
          },
          "type": {
            "type": "string",
            "example": "BetPlaced",
            "enum": [
              "BetPlaced",
              "BetWon",
              "BetLost",
              "BetVoided"
            ]
          },
          "occurredAt": {
            "type": "string",
            "example": "2026-06-16T10:00:00.000Z"
          }
        },
        "required": [
          "seq",
          "type",
          "occurredAt"
        ]
      },
      "SettleMarketRequest": {
        "type": "object",
        "properties": {
          "outcomes": {
            "example": [
              "lol-finale-a",
              "lol-finale-b",
              "lol-finale-draw"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "winningOutcomeId": {
            "type": "string",
            "nullable": true,
            "example": "lol-finale-a"
          },
          "voided": {
            "type": "boolean",
            "example": false,
            "description": "true = annulation (remboursement)"
          },
          "strategyKey": {
            "type": "string",
            "example": "WINNING_OUTCOME"
          }
        },
        "required": [
          "outcomes"
        ]
      },
      "SettleMarketResultDto": {
        "type": "object",
        "properties": {
          "settled": {
            "type": "number",
            "example": 2
          },
          "failed": {
            "type": "number",
            "example": 0
          },
          "won": {
            "type": "number",
            "example": 1
          },
          "lost": {
            "type": "number",
            "example": 1
          },
          "voided": {
            "type": "number",
            "example": 0
          },
          "failedBetIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "settled",
          "failed",
          "won",
          "lost",
          "voided",
          "failedBetIds"
        ]
      },
...
      "OutcomeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "lol-finale-a"
          },
          "label": {
            "type": "string",
            "example": "Victoire Team A"
          }
        },
        "required": [
          "id",
          "label"
        ]
      },
      "MarketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "mkt-lol-finale"
          },
          "name": {
            "type": "string",
            "example": "BetNext Major — Team A vs Team B"
          },
          "game": {
            "type": "string",
            "example": "LoL"
          },
          "outcomes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutcomeDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "game",
          "outcomes"
        ]
      },
      "CreateMarketRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "CS Major — NaVi vs Vitality"
          },
          "game": {
            "type": "string",
            "example": "CS2"
          },
          "outcomes": {
            "example": [
              "Victoire NaVi",
              "Victoire Vitality"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "game",
          "outcomes"
        ]
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "demo-player"
          },
          "password": {
            "type": "string",
            "example": "changeme123",
            "minLength": 8
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RegisterResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3"
          },
          "username": {
            "type": "string",
            "example": "demo-player"
          },
          "role": {
            "type": "string",
            "example": "PLAYER"
          }
        },
        "required": [
          "id",
          "username",
          "role"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "demo-player"
          },
          "password": {
            "type": "string",
            "example": "changeme123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginResultDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "a1b2c3"
          },
          "role": {
            "type": "string",
            "example": "PLAYER"
          },
          "token": {
            "type": "string",
            "description": "Token Bearer à envoyer dans Authorization"
          },
          "expiresInSec": {
            "type": "number",
            "example": 3600
          }
        },
        "required": [
          "userId",
          "role",
          "token",
          "expiresInSec"
        ]
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "a1b2c3"
          },
          "role": {
            "type": "string",
            "example": "PLAYER"
          }
        },
        "required": [
          "userId",
          "role"
        ]
      }
    }
  }
}